requirements.txt
pytorch = 1.3.0
torchvision = 0.4.1
numpy = 1.17.2
pillow = 6.2.0
